modify create resource methods  and

+++ src/main/java/com/uvaXplore/entity/Resource.java
@@ -13,7 +13,7 @@ public class Resource {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int resourceId;
+    private Long resourceId;
 
     private String title;
